refactor. BookingController 분리
