+++ BE/src/main/java/team07/airbnb/controller/booking/BookingUserController.java
@@ -1,4 +1,4 @@
-package team07.airbnb.controller;
+package team07.airbnb.controller.booking;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -20,13 +20,11 @@ import team07.airbnb.data.booking.dto.request.CreateBookingRequest;
 import team07.airbnb.data.booking.dto.response.BookingCancelResponse;
 import team07.airbnb.data.booking.dto.response.BookingCreateResponse;
 import team07.airbnb.data.booking.dto.response.BookingDetailResponse;
-import team07.airbnb.data.booking.dto.transfer.BookingInfoForPriceInfo;
 import team07.airbnb.data.booking.dto.transfer.DateInfo;
 import team07.airbnb.data.booking.dto.transfer.DistanceInfo;
 import team07.airbnb.data.booking.enums.CheckAuthType;
 import team07.airbnb.data.user.dto.response.TokenUserInfo;
 import team07.airbnb.entity.UserEntity;
-import team07.airbnb.exception.auth.UnAuthorizedException;
 import team07.airbnb.service.booking.BookingAuthService;
 import team07.airbnb.service.booking.BookingManageService;
 import team07.airbnb.service.booking.BookingInquiryService;
@@ -45,7 +43,7 @@ import static team07.airbnb.data.user.enums.Role.*;
 @RequestMapping("/booking")
 @RequiredArgsConstructor
 @Slf4j
-public class BookingController {
+public class BookingUserController {
 
     private final BookingInquiryService bookingInquiryService;
     private final UserService userService;
@@ -53,18 +51,6 @@ public class BookingController {
     private final BookingManageService bookingManageService;
     private final BookingAuthService bookingAuthService;
 
-    @Tag(name = "User")
-    @Operation(summary = "예약 요금 정보 조회", description = "예약의 요금 정보를 조회합니다.")
-    @GetMapping
-    @ResponseStatus(OK)
-    public PriceInfo getBookingPriceInfo(@RequestBody @Valid BookingRequest requestInfo) {
-        return bookingPriceService.getPriceInfo(
-                requestInfo.accommodationId(),
-                requestInfo.checkIn(),
-                requestInfo.checkOut(),
-                requestInfo.headCount()
-        );
-    }
 
     @Tag(name = "User")
     @Operation(summary = "예약 생성", description = "상품을 예약합니다.")
@@ -73,18 +59,6 @@ public class BookingController {
     @ResponseStatus(CREATED)
     public BookingCreateResponse createBookingRequest(@RequestBody @Valid CreateBookingRequest request, TokenUserInfo user) {
         return bookingManageService.createBooking(request, userService.getCompleteUser(user));
-    }
-
-    @Tag(name = "Host")
-    @Operation(summary = "예약 확정", description = "예약 상품의 호스트가 예약을 확정합니다.")
-    @PostMapping("/confirm/{bookingId}")
-    @Authenticated(HOST)
-    @ResponseStatus(OK)
-    public Long confirmBooking(@PathVariable Long bookingId, TokenUserInfo user) {
-        UserEntity host = bookingAuthService.currentUserIsSameWith(bookingId, user, CheckAuthType.HOST);
-
-        //컨펌한 예약의 아이디 리턴
-        return bookingManageService.confirmBooking(bookingId, host);
     }
 
     @Tag(name = "User")
@@ -101,16 +75,6 @@ public class BookingController {
         return BookingCancelResponse.of(bookingId, cancelFee);
     }
 
-    @Tag(name = "Host")
-    @Operation(summary = "예약 이용 완료", description = "예약을 이용 완료 처리합니다.")
-    @PostMapping("/complete/{bookingId}")
-    @Authenticated(HOST)
-    public void completeBooking(@PathVariable Long bookingId, TokenUserInfo user) {
-        UserEntity host = bookingAuthService.currentUserIsSameWith(bookingId, user, CheckAuthType.HOST);
-
-        // 예약 종료 일자 전 예약 이용 완료 -> 남은 일자에 대해서 상품 재생성, 환불 X
-        bookingManageService.reopenBooking(bookingId);
-    }
 
     @Tag(name = "User")
     @Operation(summary = "내 예약 조회", description = "내 예약을 조회합니다.")
@@ -128,37 +92,5 @@ public class BookingController {
         bookingAuthService.currentUserIsSameWith(bookingId, userInfo, CheckAuthType.USER);
 
         return BookingDetailResponse.of(bookingInquiryService.findByBookingId(bookingId));
-    }
-
-
-    @Tag(name = "Host")
-    @Operation(summary = "내 숙소의 예약 조회", description = "내가 등록한 숙소의 예약을 조회합니다.")
-    @Authenticated(HOST)
-    @GetMapping("/management")
-    @ResponseStatus(OK)
-    public List<BookingDetailResponse> getBookingInfosOfHosting(TokenUserInfo host) {
-        return bookingInquiryService.getBookingInfoListByHost(userService.getCompleteUser(host));
-    }
-
-
-    @Operation(summary = "호스트 예약 상세 조회", description = "호스트가 예약 상세 정보를 조회합니다.")
-    @GetMapping("/management/{bookingId}")
-    @Authenticated(HOST)
-    @ResponseStatus(OK)
-    public BookingDetailResponse getBookingDetail(@PathVariable Long bookingId, TokenUserInfo host) {
-        bookingAuthService.currentUserIsSameWith(bookingId, host, CheckAuthType.HOST);
-        return BookingDetailResponse.of(bookingInquiryService.findByBookingId(bookingId));
-    }
-
-    @Operation(summary = "날짜와 지역에 따른 요금정보들 조회", description = "요금 그래프를 위한 선택한 지역, 날짜의 평균 요금들을 정렬된 리스트로 반환합니다.")
-    @GetMapping("/pay-info")
-    @ResponseStatus(OK)
-    public List<Double> getPayInfo(@RequestBody BookingPaymentsRequest bookingPaymentsRequest) {
-        List<Double> result = bookingInquiryService.getPricesAccAvailable(
-                                DateInfo.of(bookingPaymentsRequest),
-                                DistanceInfo.of(bookingPaymentsRequest)
-                        );
-        result.sort(Comparator.naturalOrder());
-        return result;
     }
 }
